Remove Pipenv and update dependencies
Pipenv is causing problems with dependency locking and fixing the python
version to 3.7. Fixing these issues are easier with plain old pip and
given that Pipenv isn't designed for managing dependencies with packages
this seems effort that is not worth it.

This change removes Pipenv, the docker stuff (which wasn't used anyway),
updates the dependencies in a couple of requirements files and fixes the
upgrade issue with flask_marshmallow